textrenderer types

+++ edbee-lib/edbee/views/components/texteditorrenderer.h
@@ -16,18 +16,18 @@ class TextTheme;
 
 class EDBEE_EXPORT TextEditorRenderer {
 public:
-    TextEditorRenderer( TextRenderer *renderer );
+    TextEditorRenderer(TextRenderer *renderer);
     virtual ~TextEditorRenderer();
 
     virtual int preferedWidth();
     virtual void render(QPainter* painter);
-    virtual void renderLineBackground(QPainter *painter, int line);
-    virtual void renderLineSelection(QPainter *painter, int line);
-    virtual void renderLineBorderedRanges(QPainter *painter, int line);
-    virtual void renderLineSeparator(QPainter *painter, int line);
-    virtual void renderLineText(QPainter *painter, int line);
+    virtual void renderLineBackground(QPainter *painter, size_t line);
+    virtual void renderLineSelection(QPainter *painter, size_t line);
+    virtual void renderLineBorderedRanges(QPainter *painter, size_t line);
+    virtual void renderLineSeparator(QPainter *painter, size_t line);
+    virtual void renderLineText(QPainter *painter, size_t line);
     virtual void renderCarets(QPainter *painter);
-    virtual void renderShade(QPainter *painter, const QRect& rect );
+    virtual void renderShade(QPainter *painter, const QRect& rect);
 
     virtual int extraPixelsToUpdateAroundLines();
 
@@ -37,7 +37,6 @@ private:
     TextRenderer* rendererRef_;       ///< the renderere reference
     TextTheme* themeRef_;             ///< A theem reference used while rendering
     QLinearGradient* shadowGradient_; ///< The shadow gradient to draw
-
 };
 
 } // edbee
